Add .cfignore
This is a copy of .gitignore, but with yarn.lock added.

It is unclear how or why Yarn is used alongside NPM.

If yarn.lock is pushed you receive the error:

```
Lockfile does not contain pattern "cfenv@^1.1.0"
```

Presumably because I would have had to recreate yarn.lock after adding
cfenv to the package.json. However, I did not do this because for Cloud
Foundry deployments the dependencies and compilation steps happen during
the staging process, so lock files are not required.